add tests for ChangeService.
add check for existence of a change to updateChangeStatus and updateChangeProductId of ChangeService.

change return value of updateChangeStatus and updateChangeProductId to Change.

add NoSuchChangeException exception.

add NoSuchChangeException-throw to updateChangeStatus and updateChangeProductId in ChangeService

add handleNoSuchChangeException to GlobalExceptionHandler giving an INTERNAL_SERVER_ERROR, since a user shouldn't be able to look at a single change just yet.